chore: renamed schemaStore files to schemaDeploy
