+++ packages/sanity/src/_internal/cli/actions/schema/utils/mainfestExtractor.ts
@@ -2,7 +2,7 @@ import {type CliCommandArguments, type CliCommandContext, type CliOutputter} fro
 import chalk from 'chalk'
 
 import {type ExtractManifestFlags, extractManifestSafe} from '../../manifest/extractManifestAction'
-import {FlagValidationError} from './schemaStoreValidation'
+import {FlagValidationError} from './schemaDeployValidation'
 
 export type ManifestExtractor = (manifestDir: string) => Promise<void>
