+++ packages/sanity/src/_internal/cli/commands/index.ts
@@ -1,6 +1,6 @@
 import {type CliCommandDefinition, type CliCommandGroupDefinition} from '@sanity/cli'
 
-import {SCHEMA_STORE_FEATURE_ENABLED} from '../actions/schema/schemaStoreConstants'
+import {SCHEMA_STORE_FEATURE_ENABLED} from '../actions/schema/schemaDeployConstants'
 import backupGroup from './backup/backupGroup'
 import disableBackupCommand from './backup/disableBackupCommand'
 import downloadBackupCommand from './backup/downloadBackupCommand'
