+++ packages/sanity/test/cli/schema/deleteSchemasAction.test.ts
@@ -7,15 +7,15 @@ import {
   deleteSchemaAction,
   type DeleteSchemaFlags,
 } from '../../../src/_internal/cli/actions/schema/deleteSchemaAction'
-import {type SchemaStoreContext} from '../../../src/_internal/cli/actions/schema/schemaStoreTypes'
+import {type SchemaStoreContext} from '../../../src/_internal/cli/actions/schema/schemaDeployTypes'
 import {getWorkspaceSchemaId} from '../../../src/_internal/cli/actions/schema/utils/workspaceSchemaId'
 import {type StoredWorkspaceSchema} from '../../../src/_internal/manifest/manifestTypes'
-import {createSchemaStoreFixture} from './mocks/schemaStoreFixture'
+import {createSchemaStoreFixture} from './mocks/schemaDeployFixture'
 import {
   createMockJsonReader,
   createMockSanityClient,
   createMockStoreSchemaContext,
-} from './mocks/schemaStoreMocks'
+} from './mocks/schemaDeployMocks'
 
 const fixture = createSchemaStoreFixture(new Date().toISOString())
 const {files, testSchema, testWorkspace, testMultiWorkspaceManifest, staticDate, workDir} = fixture
